fixed tokenization
forgot to put the BOS and EOS tokens onto each example.  Added these and
boosted accuracy from 61% to 83%.  I was also calculating the length of
the example incorrectly which was causing the attention mask to not be
properly sized.

Also created a script to trace the model instead of putting the
instructions in the README.  Lastly, I removed the absolute paths and
made them relative paths.